aufteilen von state file
