+++ frontend/lib/states/riverpod_states.dart
@@ -1,5 +1,4 @@
 import 'package:PointCollector/repository/ProductRepository.dart';
-import 'package:PointCollector/repository/UserRepository.dart';
 import 'package:PointCollector/screens/home_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -76,15 +75,4 @@ class ProductState extends StateNotifier<Future<List<ProductModel>>> {
 final productProvider =
     StateNotifierProvider<ProductState, Future<List<ProductModel>>>((ref) {
   return ProductState();
-});
-
-class RedeemPointState extends StateNotifier<int> {
-  RedeemPointState() : super(UserRepository().getPoints());
-
-  void setPoints(int points) => state = points;
-}
-
-final redeemPointProvider =
-    StateNotifierProvider<RedeemPointState, int>((ref) {
-  return RedeemPointState();
 });
